service_market_uz/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-phonenumber-field[phonenumbers]>=7.2.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
Pillow>=10.2.0
gunicorn>=21.2.0
django-filter>=23.5
python-telegram-bot>=20.7
django-unfold>=0.18.0
